--- app/data.py
@@ -1,68 +1,0 @@
-EVENTS = [
-    {
-        'id':1,
-        'title': u'Drink Smoothie',
-        'imageUrl': u'http://freedesignfile.com/upload/2016/11/Fresh-fruit-with-smoothies-HD-picture.jpg',
-        'location': u'Machakos, Kenya',
-        'time': u'11:00AM',
-        'date': u'25 NOV 2017',
-        'description': u'Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s,',
-        'done': False,
-        'rsvp': [
-            {
-                'user_id': 1,
-                'name': u'John Doe',
-                'email': u'john.D@gmail.com'
-            },
-            {
-                'user_id': 3,
-                'name': u'Antony Ng\'ang\'a',
-                'email': u'tonny.nesh@gmail.com'
-            }
-        ]
-    },
-    {
-        'id':2,
-        'title': u'Enviroment Awareness',
-        'imageUrl': u'https://s3-us-west-1.amazonaws.com/stanford.ucomm.newsms.media/wp-content/uploads/2016/11/10165436/environment_GettyImages-501231894.jpg',
-        'location':u'Karura Forest, Kenya',
-        'time':u'07:00PM',
-        'date':u'30 NOV 2017',
-        'description': u'Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s,',
-        'done': False,
-        'rsvp': [
-            {
-                'user_id': 2,
-                'name': u'Mary Jane',
-                'email': u'jane.mary@yahoo.com'
-            }
-        ]
-    },
-    {
-        'id':3,
-        'title': u'Tattoo Fest',
-        'imageUrl': u'http://ww1.prweb.com/prfiles/2017/09/29/14755349/Ray-Price-Capital-City-Bikefest_TattooFest_7574.jpg',
-        'location':u'Nairobi, Kenya',
-        'time':u'03:00PM',
-        'date':u'20 NOV 2017',
-        'description': u'Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s,',
-        'done': False,
-        'rsvp': [
-            {
-                'user_id': 4,
-                'name': u'Mary Jane',
-                'email': u'jane.mary@yahoo.com'
-            },
-            {
-                'user_id': 1,
-                'name': u'James Micheal',
-                'email': u'micheal.j@gmail.com'
-            },
-            {
-                'user_id': 5,
-                'name': u'Francis Gwada',
-                'email': u'gwada.frank@mango.com'
-            }
-        ]
-    }
-]
