--- app/user_data.py
@@ -1,20 +1,0 @@
-USERS = [
-    {
-        'id':1,
-        'name': u'John Doe',
-        'email': u'john.D@gmail.com',
-        'password': u'qwerty1234'
-    },
-    {
-        'id':2,
-        'name': u'Mary Jane',
-        'email': u'jane.mary@yahoo.com',
-        'password': u'qwerty1234'
-    },
-    {
-        'id':3,
-        'name': u'Antony Ng\'ang\'a',
-        'email': u'tonny.nesh@gmail.com',
-        'password': u'qwerty1234'
-    }
-]
